Exclude drafts from live pagelist
- Exclude drafts from live pagelist
- Extract controller into standalone file
